+++ test/test_helper.bash
@@ -7,6 +7,9 @@ setup() {
   # Use BATS provided temporary directory
   cd "$BATS_TEST_TMPDIR"
 
+  # Override XDG_DATA_HOME to use temporary directory for tests
+  export XDG_DATA_HOME="$BATS_TEST_TMPDIR/.xdg-data"
+
   JOB_CLI="$BATS_TEST_DIRNAME/../dist/gob"
 }